+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/IGameSessionService.cs
@@ -20,8 +20,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Runtime.CompilerServices;
+using System.Threading;
 using System.Threading.Tasks;
+using Newtonsoft.Json.Linq;
 using Stormancer.Server.Plugins.GameSession.Models;
 
 namespace Stormancer.Server.Plugins.GameSession
@@ -40,6 +44,21 @@ namespace Stormancer.Server.Plugins.GameSession
 
         bool IsHost(string sessionId);
 
-     
+        /// <summary>
+        /// Make this game session open to new players via the specified <paramref name="gameFinder"/>.
+        /// </summary>
+        /// <param name="data">Custom data that will be passed to <paramref name="gameFinder"/>.</param>
+        /// <param name="gameFinder">Name of the GameFinder where the session will be made available.</param>
+        /// <param name="ct">When this token is canceled, the game session will be withdrawn from <paramref name="gameFinder"/>.</param>
+        /// <returns>
+        /// An async enumerable that yields every Team that is added to the game session by <paramref name="gameFinder"/>.
+        /// It completes when the game session has been withdrawn from <paramref name="gameFinder"/>.
+        /// </returns>
+        /// <remarks>
+        /// When this method is called, the game session will be made available to the GameFinder designated by <paramref name="gameFinder"/>.
+        /// From there, players in <paramref name="gameFinder"/> will be able to join it, provided <paramref name="gameFinder"/> implements the required logic.
+        /// The game session stays open until <paramref name="ct"/> is canceled.
+        /// </remarks>
+        IAsyncEnumerable<Team> OpenToGameFinder(JObject data, string gameFinder, CancellationToken ct);
     }
 }