+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -17,3 +17,4 @@ Added
 
 - ``GameSessionConfigurationDto`` now has an additional ``HostUserId`` member. This member will be set in the object returned by ``GameSessionService.GetGameSessionConfig()`` if the game session has P2P enabled.
 - ``IGameSessionEventHandler``: new ``OnClientConnected()`` method.
+- New ``IGameSessionService.OpenToGameFinder()`` method. Allows adding new players to the session after it has started.